+++ configuration.go
@@ -2,10 +2,6 @@
 
 package webrtc
 
-import (
-	"github.com/pion/ice"
-)
-
 // Configuration defines a set of parameters to configure how the
 // peer-to-peer communication via PeerConnection is established or
 // re-established.
@@ -50,18 +46,4 @@ type Configuration struct {
 	// SDPSemantics controls the type of SDP offers accepted by and
 	// SDP answers generated by the PeerConnection.
 	SDPSemantics SDPSemantics
-}
-
-func (c Configuration) getICEServers() (*[]*ice.URL, error) {
-	var iceServers []*ice.URL
-	for _, server := range c.ICEServers {
-		for _, rawURL := range server.URLs {
-			url, err := ice.ParseURL(rawURL)
-			if err != nil {
-				return nil, err
-			}
-			iceServers = append(iceServers, url)
-		}
-	}
-	return &iceServers, nil
 }
